Admin panel users page: deal with corrupted DBs where accounts reference missing users
On at least alpha.sandstorm.io and zero (@kentonv's dev instance), there are
user accounts in Mongo's users collection with a loginIdentities field that
references an identity account for which there is no document in the
collection.

Theoretically this is supposed to be impossible, since we only ever delete
identities that are demo identities, and at least alpha has never enabled demo
accounts.  In practice, these DBs exist, and we'd do well to handle their
(surprising) contents a little more defensively.