accounts-email-token: save/restore path in email login flow
One problem with the email login workflow is that clicking the link in the
email loses the URL you were at before starting the login flow (which in turn,
might have referenced a sharing token or a grain ID).

This patch modifies accounts-email-token and sandstorm-accounts-ui to:

* Send the current path to the server when requesting an email login token
* Saves the login-initiating path in the DB with the generated token
* Passes the saved path back to the client on successful email token login
* Redirect to the login-initiating path on completion of login

This helps smooth out the email login flow - the tab opened by following the
link is no longer inferior to the tab you initiated email login from.