Fix large downloads from apps buffering in memory.
This was causing several processes to allocate excessive memory as a result of any large download:
- sandstorm-http-bridge
- the backend coordinator
- the front-end

This change also reduces the total number of write() calls performed which improves performance. Previously, a write() RPC was occurring for every "chunk" in the HTTP chunking sense.

Unfortunately, an app update is required. But, only a few apps do large downloads.

Fixes #1982
Fixes #1054