run-local.sh: Switch back to using fixed port assignments
We use ports 9000-9004 now.  This was chosen to not conflict with 6080, our
usual test port, or anything else in particular.

The hope is that this patch will make the email sending tests stop hitting
EADDRINUSE.  Even if they don't, we can consider using network namespaces
to make sure nothing else is messing with the tests.