Proactively end streaming when ResponseStream.end() is called.
There is currently a cycle created between sandstorm-http-bridge and proxy.js where proxy.js doesn't drop the stream handle until the ResponseStream is dropped, but the ResponseStream is not dropped until the handle is dropped.

By having proxy.js drop the handle when end() is called, we avoid the cycle at least for successful streams. For unsuccessful streams (that never call end()) there is still a problem, but it needs to be solved on the sandstorm-http-bridge side.