Refactor accounts-identity UI
This refactor makes explicit all of the states that the UI can be in.  We also
handle any Meteor method call failures, displaying the error message in the UI
in an accessible manner.  This will allow us to add UI to explain why login
with identity fails when a user limit is reached, and provides a reasonable
place for us to prompt for two-factor auth in the more distant future.

In the non-login-identity view, we:

* split the identity list out by account with a faint horizontal rule
* retain only one "unlink" button per account, since unlinking is an action
  taken on the account, rather than on the identity pictured
* avoid using window.confirm() in favor of an in-app modal dialog
* visibly handle success/failure of any of the Meteor method calls with
  the focusing success/error boxes

Additional stylistic changes include avoiding using the `event` global,
using evt.currentTarget rather than evt.target, and preferring ES6 fat-arrow
functions in most places.