Rework profile picture upload.
We now also get error/success messages on the "Confirm your profile" page,
and we don't have to alert() on failures any more.

Additionally, we clean up the upload lifecycle to prevent a number of weird
corner cases the browser could get into.  For instance, if you

1. clicked the upload button
2. chose no file
3. clicked the button again
4. and chose a file

Then the browser would upload that same file twice, and it'd be a race.

Or if in step 4 you chose a nonexistent file, you'd get two window.alert()s.