Update npm shrinkwrap
When we added the heapdump dependency via NPM, we also needed to update the
shrinkwrap to specify precisely which version.

Updating the shrinkwrap was nontrivial because npm3 has a bug.  In particular,
npm3 (which we got upgraded to by the recent Meteor upgrade to 1.3.3) appears
to have a bug in resolving dependencies from git that specify a branch name
rather than a SHA1: https://github.com/npm/npm/issues/10502 .  To work around
this, we specify our intro.js dependency via the current sha1.  (We should probably
try to get back on an upstream intro.js release.)

We'd probably benefit from a lint/pre-commit hook that checks that commits that
modify `package.json` also modify `npm-shrinkwrap.json`.