Route/proxy non-main-host requests before sockjs's handler swallows them
This fixes the longstanding "Meteor app refresh loop" issue.

Meteor's sockjs inserts itself as the first request listener on the Meteor HTTP
handler object.  Normally, this is fine, and keeps DDP requests from
accidentally getting routed to the application handlers.  This is problematic
for programs that happen to implement a proxy, and which need to proxy requests
that use the same paths and overall protocol as Meteor does internally.

As a result of the sockjs middleware being first in the chain, XHRs to an
Meteor app's sockjs endpoint would wind up getting answered by Sandstorm's
instance of sockjs and communicating over DDP to Sandstorm itself.  So any app
that fell back to long-polling would speak DDP to Sandstorm, rather than to the
intended app's backend.

Meteor hashes all the code that goes into an app to determine if a reload is
needed to have the latest code running on the client.  It publishes the current
version of the client code in a collection over DDP.  This lead the code
running in the browser to believe that there had been a code change, since it
was now getting the hash for **Sandstorm's** shell code over DDP, which did not
match the hash for the app's code.  So the app would refresh the page, hoping
to load the new code.  Browsers would cache that websockets had failed for that
host, and so the app would load the code, learn about Sandstorm's code version
over sockjs long-polling DDP, and attempt to refresh again.

The solution here is to do Sandstorm's proxying earlier in the request dispatch
chain.  We splice our own handler in after ddp-server has installed its sockjs
handler, and now everything works as intended.