Define a capability interface for identities.
Using identity IDs in the notification interface was problematic because sandstorm-http-bridge truncates the identity IDs that it sends to the app, so the app actually doesn't have the full ID to send back to Sandstorm in the API.

I decided to go ahead and take the step of defining an Identity capability type to use instead, which is something that was going to happen at some point anyway. This capability is then what you put on the ActivityEvent. Over time, we should move away from identity IDs and towards capability representations, as this makes it easier for us to do things like remap identities when moving grains between servers, which may allow us to get away from the requirement that identity providers be global.

Of course, this doesn't really solve the problem for sandstorm-http-bridge users -- now they need a way to get this capability. So I extended sandstorm-http-bridge with a new feature where it can automatically save Identity capabilities when first observed, and then give the app a way to fetch those capabilities later based on (truncated) identity ID.