Fix loading anything at all under IE10
A package was using "const" without declaring api.use("ecmascript")
in its package.js.  This resulted in it not getting passed through
the transpilation pipeline, and having its contents placed in the
bundle directly.

IE10 does not support the "const" keyword, and parse errors when it
sees this code, and stops attempting to parse JS.