Give top bar a z-index of 0
Using 0 rather than CSS default ("auto") allows introjs to highlight elements in the topbar.
Interestingly, at runtime, Chrome reports the z-index of the topbar element as "0" before
this change; Firefox reports it as "auto." By setting it to 0 explicitly, we cause Firefox
and Chrome to have the same behavior when queried by introjs about the top bar's z-index.

Close #2163