Move email-related things into package.json and shell/
We are currently shipping two separate copies of simplesmtp and mailcomposer
for sandstorm-email and for the email driver.  This is probably unwise.

This patchset canonicalizes on the more recent of each package depended on by
either package.  It is likely that we will want to upgrade these dependencies
Soon™, which will become possible once we upgrade to Meteor 1.4.