Fix invalid SCSS (@extend and @media don't mix)
Due to how @extend is implemented, it can't be placed in a block that is a
child of a media query.  @mixin has no such restriction.

We previously used @extend to produce smaller .css files (though in practice
gzip winds up making the two files basically the same size on the wire).  This
patch switches the two partials we use inside media query blocks (specifically,
unstyled-button and overflow-ellipsis) to be mixins instead.