Use a version of simplesmtp with a proper "main" directive
The deprecated simplesmtp has a package.json which specifies:

    "main": "./lib/smtp",

but no such file exists.  In fact, the appropriate entry point would be:

    "main": "./index.js",

I've applied that change to a fork of simplesmtp until we're able to move on,
because somehow the invalid "main" directive doesn't break the shell when run
in dev mode, but Meteor seems to believe the package doesn't exist at all when
running from the hermetic bundle.