pseudo-img-tag should use contain, not cover
`background-size: cover` and `background-size: contain` are related, but distinct:

`contain` means "scale the image preserving aspect ratio until the image is
perfectly contained in the size provided, and letterbox center it if the image
is too small in one dimension for the bounding box".

`cover` means "scale the image preserving aspect ratio until it completely
covers both the horizontal and vertical space of the bounding box, then clip
whatever overflowed, if the image is too big in one dimension."

For images with an intrinsic aspect ratio that matches that of the containing
box, the two are equivalent.  This is true of all of our internal usage with
our own assets.

For images that may not match the bounding box's aspect ratio, contain is much
more likely to produce the desired result, which is that the image drawn fits
neatly in the containing box but is a little smaller, rather than clipping
arbitrary pieces of the image and having what remains be unusually large.

So this patchset switches the behavior of %pseudo-img-tag to use contain rather
than cover.

Refs #2241.