Refactor/prettify notification templates
This pulls apart the six types of notifications into separate templates
and helpers, which makes it much easier to modify each without.

All notifications leave space for an icon, because having the notification text
aligned to a consistent grid makes the design look much cleaner.  If we want to
add icons for the referral program and the request to send us statistics, we
can do that pretty easily in the appropriate templates.

We also switch to using the standard-form styling for buttons, and the
default-action-on-right style.