Reimplement email support atop nodemailer
`simplesmtp` is unmaintained and has major problems.  We are dropping it.

`nodemailer` (by the same author) is a collection of packages which together
form the most widely used JS mail library.  It provides a number of transports,
including speaking SMTP directly, to actually deliver mail.  It's possible
we'll want to support transports other than just plain SMTP; this change paves
the way to do so neatly.  We use the `nodemailer-smtp-pool` transport to
deliver mail, which is comparable to the `simplesmtp` pool.

`mailcomposer` is an implementation detail of nodemailer, so we no longer need
to pull it in explicitly.  Mail is expressed as simply a plain JS object.

`mimelib` is no longer used, so I have removed it from our package.json.

`smtp-server` is a package which replaces simplesmtp's server support.  It was
pretty simple to drop in.  We configure it to not require authentication,
because we can't require auth for receiving email.

`mailparser` (used by the email driver to parse inbound email) has changed its
API a bit to separate envelope session information from message data, so we
change how we process mail in the inbound mail driver a bit.

I have manually tested:

* Email self-test in setup/admin panel
* Email sent by Roundcube gets delivered correctly (modulo the extra trailing .
  that the sandstorm-smtp-bridge inside the app package injects for text/plain mail)
* Email sent to grain gets delivered to Roundcube

And this should pass the testsuite.