Change how Notifications compute client-side titles
The publication of just the grain title in Grains is surprising, and produces
the wrong title if you've set a name for your ApiToken.

This change removes that publication, extracts the code for picking a title
when given an ApiToken into a helper for reuse, and uses said code in the
various places we need to pick a title, including GrainView, notifications,
and desktop notifications.