Don't check that mail server certs are signed by any CA in particular
This is restoring the behavior that existed before #2362.  Several users have
reported that the most recent upgrade has broken mail delivery on their systems
due to attempting to perform STARTTLS, then aborting when the cert provided by
the mailserver didn't validate.  This appears to be a fairly common configuration.

Node 0.10 did not validate certs at all before now, so I'm disabling cert
validation to undo the breakage.  Longer-term, we should add UI for admins to
specify one of the following behaviors:

* "none" use STARTTLS if available, but don't validate the cert
* "starttls" require STARTTLS, and validate the certificate
* "smtps" connect on port 465 and validate the cert