Inject db into migrations
This changeset extracts migrations from the sandstorm-db package, allowing the
database to be injected as an argument, rather than relying on a broad swath of
globals.

In doing so, some other accounts-specific code that previously had to live in
sandstorm-db (so migrations could call it safely) can move out to accounts,
where it belongs.