Avoid using variables named "event" throughout the codebase.
An unfortunate piece of legacy behavior from IE is the `window.event` global
variable, accessible via "event".  The precise behavior of this global varies
between the browsers, and it's easy to have code that accidentally accesses the
global when it means to access a local that is shadowing the global.

For cleanliness, prefer to use a different name (I like `evt` and `ev`) for
variables you might otherwise name `event`.  This way you can't accidentally
write code that reads the global and works in Chrome and IE but not Firefox.

In particular, depending on window.event caused the "Unsubscribe" button to
break on Oasis.  I filed the appropriate fix against blackrock, but it seemed
worthwhile to ensure we're not making the same mistake elsewhere.