Audit for Blaze implicit data context propagation
It is difficult at a glance to tell if Blaze template inclusions like
`{{> templateName }}` are intended to be simple includes, or if they are also
supposed to receive everything that is in the current data context as well.

This change clears up the confusion by never writing a bare `{{> templateName }}`:
instead, in places where we can safely pass on no data context, we write
`{{> templateName "" }}`, and in places where we must still pass on the entire
data context, we write `{{> templateName . }}` to be explicit about it.