Make danger buttons not so loud
The white-on-pure-red is somewhat low-contrast.  The saturation of the
red is also very eye-catching and attention-grabbing, when most danger
buttons are not a part of the UI that we wish to emphasize.

This change demphasizes danger buttons, but preserves the big red background on
hover and keeps the text red to indicate that the action is destructive to e.g.
mobile users that won't see hover states.

Fixes #2702