Build and ship a patched node in the Sandstorm bundle
v8's performance in the face of many fibers is currently quite poor because
they use a linked list for all thread control block lookups.  We've observed
frontends on Oasis sitting with 65% CPU usage just navigating that linked
list.  See #2484 for screenshots.

Switching to a real data structure (a hash map) improves this situation
drastically.  @jparyani's test app [1] spends 46% of CPU time in
Isolate::FindOrAllocatePerThreadDataForThisThread when run with the default
meteor-bundled node.  That function doesn't even register in the perf report
when run against a node compiled with the hashmap patch.

I'm not thrilled about building node and carrying this patch, so if we can get
a proper fix upstream, that would be preferable.

[1] - https://github.com/jparyani/meteor-fiber-repro