Harden against CSRF attacks on apps.
Chrome and (allegedly) Safari send an "Origin" header on *all* non-GET/HEAD requests. Thus, we can totally block CSRF attacks on apps from these browsers by rejecting requests with an unexpected Origin header.

For other browsers, we might be able to block CSRF attacks based on the Referer header. Of course, it's easy for an attacker to suppress the Referer header, so in order to create more than a minor roadbump for them, we'd need to reject requests that lack a Referer header in addition to those with a mismatching header.

Unfortunately, there are a small number of users who intentionally configure their browser to suppress the Referer header in all cases. Probably, these users are rightly concerned about the possibility of cross-site information leaks and privacy violations that the Referer header can create, and would not be concerned with same-origin Referer headers (which is what we need), but unfortunately the configuration option provided by Firefox does not distinguish between these.

For people who *don't* block the Referer header, it seems like we can implement meaningful CSRF protection. Since this is the vast majority of users, it seems worth implementing. But we don't want to break the users who do block it. So, what we can do is keep track of whether we've ever seen a Referer header on a particular session and, if so, assume that Referer headers are enabled and require their presence on future non-GET requests.

I am not 100% confident in this assesment, so for now I have only implemented some logging. Once this is running in Oasis for a few days, we can grep the logs to see which cases occur in practice.