If the user tries to set an expired key, automatically check for a newer version.
We don't attempt to renew it, we just check if it had been renewed before.

If no current key can be found, don't let the user install an expired key. Once a key is installed, we may auto-renew it, but we don't want a user to accidentally upload an expired key and then be charged for renewal unexpectedly.