Soften our claim about being the only way to make open source web apps viable.
The intended meaning was: "Making it easy to run apps on your own server is the only way to make open source web apps viable."

However, it was easy to interpret: "Sandstorm is the only way to make open source web apps viable."

I've removed the word "only".

Additionally, Sacha Greif on Twitter expressed that the word "viable" here seems confusing, so I found a way to express the same thing without that word -- though of course the blog post still uses the word a lot.