Fix that the preinstalled app package IDs could become outdated.
When an app update is discovered through an app index refresh, if the newer version of the app was already installed (for example, because someone had gone to the app market and installed it manually before the index update landed), then the preinstall package ID was never updated. This means the old version continued to be preinstalled and also broke appdemos of the particular app.