Don't escape HTML sharing e-mails client-side.
This isn't actually a security problem because we don't render this HTML in our origin -- we send it in an e-mail. Moreover, the e-mail has the initiating user's address as the "from" address. The user could just as easily have sent any malicious e-mail content directly.

However, it looks really suspicious for us to be doing this client-side, so let's not.