Don't count trashed grains against grain quota.
Users find this confusing.

A better solution would be to automatically delete trashed grains as needed when the user is running up against quota limits, but that would be trickier to implement, and as a result we haven't done it. In the meantime, this stopgap is much better than nothing.

Technically a user could abuse this by swapping grains in and out of their trash, but whatever.