Allow apps to display 'data:' and 'blob:' URLs in their iframe.
The purposes of restricting frame-src is more to prevent weird behavior than it is security: We don't want the grain iframe to browse to an external site because that would be weird and confusing. I don't see any reason to prevent data or blob URLs, though, and they are very useful for some kinds of apps.

Fixes #2602