UiView powerbox: sort unopened apiTokens by created if no lastUsed
Jade reported that she struggled to find her recently-created item to attach
when doing a powerbox query, and that the grains seemed unordered.

While we do sort the grains by last used time, it seems that grains shared with
you that you never opened sit at the top of the list, whereas you probably want
them to show up in the grain list by their creation timestamp, as we do in the
grain list.