Make IpInterface test reliable
The IpInterface tests have been flaky since their incept due to a race
condition between the testsuite attempting to connect to the port the test
grain requests to listenTcp on and the grain actually getting the frontend to
bind/listen that port.

The changes to the test app in [1] make a visible-on-the-page change when the
server acknowledges having completed binding the port, so we can now wait for
that before calling `assertTcpConnection()`, which in turn should prevent this
race condition from causing spurious test failures on Jenkins.

[1] https://github.com/sandstorm-io/sandstorm-test-python/pull/7