Update isKernelNewEnough() to require 3.10 instead of 3.13.
We choose 3.10 specifically because it's what RHEL / CentOS 7 uses. It's possible we could go back as far as 3.5 (when seccomp was introduced), but this is not tested.

Stop checking the userns sysctl because we're actually going to stop requiring it.

Change the warning to an error. It should have become an error years ago, but we never bothered.