Fix running in not-started-as-root mode.
Apparently, when Sandstorm's outer container uses userns (because it was not started at root), we've been mapping the numeric UID within that container as zero. This didn't actually provide root privileges because it's in a user namespace and because we explicitly drop all capabilities before starting node and mongo. However, using the numeric value zero now trips an assert in the supervisor. It's also not a good idea to use zero when not really running as root or with root-like privileges becaues it can confuse apps.

AFAICT there was never any reason to use zero here instead of 1000, and using 1000 is actually a better match with what happens when starting as root.