installer-tests: Use a more specific cleanup function
Earlier, I had added purging postfix to cleanup for all tests. However, not all of our test VMs run Debian,
so this was a pretty silly idea in hindsight.

This Sandstorm commit uses a new cleanup function defined here:
https://github.com/paulproteus/stodgy-tester/commit/51ed3de778ca3aa289ece69024a576449a3afbe4
so that Postfix is successfully removed from the VM to avoid disruption to
other installer-tests.