Move info/error messages in the login dialog to the top of the form
Placing errors below the thing the user is interacting with is bad because:

* the errors can wind up completely off-screen if they appear below the fold
* users are less likely to associate the error message with the action they
  just took if it is in a section that they haven't read/interacted with on the
  page.

It is better to place errors at the top of the context to which they apply
than to place them below the context to which they apply.

Ideally, we'd place errors associated with a particular login mechanism just
above *that login mechanism's login section* but I don't think we currently
track those separately and this seems like a strict improvement.