Feature key upload form: show example feature key
In user testing, it took a user longer than I expected to realize what was they
were supposed to do next after obtaining a trial key.

The primary goal here is to show an example of what a feature key looks like,
so when the user obtains one they'll more quickly connect that they're supposed
to paste it into the box.  A secondary goal is to make it clear that the ascii
armor is okay, and that they can paste in the whole block, rather than
painstakingly removing the ascii armor (which we also accept, but waste's the
user's time).

If the user attempts to use the placeholder feature key, we will give them a
gentle admonishment.