Fix updates for non-root installs.
Currenly, for servers running with no root privileges, each update crashes the server. Starting the server back up again works fine.

The problem was introduced in the change introducing the privileged sandbox. On non-root installs, Sandstorm's outer container needs to use a user namespace. Within that user namespace, the UID was being mapped to zero. This caused a newly-added assert in supervisor.c++ to fail. Rather than roll back the assert, I decided to change the UID mapping to non-zero, *completely ignoring* the comment *right above* the code that explained why it really does need to be zero. :(

This change goes back to mapping as UID zero and removes the asserts instead.

Note that UID zero inside the user namespace is not actually root on the system. Also, we explicitly drop crapabilities so that the front-end process does not even get within-namespace root rights.

Unfortunately, because of the nature of the bug, the update which fixes the bug will again fail -- but updates after that will succeed.

Fixes #2692.