Setup wizard: show a sane thing when setup done but BASE_URL changed
Changing BASE_URL can deconfigure Google and GitHub login, which can leave
the server in a situation where you can't log in, but you're directed to
sign in to create an admin account.  This is senseless.

Instead, we should detect this situation and show a more useful message
explaining that you can't log in and you should reconfigure identity providers
instead.