Add in-tree test app.
This is something I've long wanted: A test app that builds directly out of the Sandstorm tree, so that it can be edited together with the tests. This eliminates the need to create a separate PR for the test app as well as the need to host the updated SPK somewhere, both of which were sufficiently burdensome that it tended to deter me from writing tests at all.