Use native setTimeout() to clean up Sandcats sockets
Meteor requires the use of Meteor.setTimeout() instead of JS' native
setTimeout() because they rely on Fibers. When you pass an async
callback to some non-Meteor function, like setTimeout(), that callback
function won't run in a Fiber, so calling any Meteor code will throw
an exception.

The problem here is that Meteor.setTimeout() counts as Meteor code
that has to be run in a Fiber, but the socket.bind() callback is
messing up the Fiber context (so ironically, the "solution" for
non-Fiber callbacks is causing crashes). One approach would be to wrap
said callback in a Fiber. However in this case, we're not calling any
Meteor code besides Meteor.setTimeout() itself, we're just cleaning up
a UDP socket. No big deal. So we use the native setTimeout().