Replace run-dev.sh with new 'sandstorm dev-shell' command.
This will automatically stop the installed shell and seamlessly switch over to the dev shell (running in the meteor dev tool) until you ctrl+c, at which point the installed shell starts back up.

This all works via FD passing. It works whether or not you have the new gateway enabled, but these changes were required for the gateway since the backend is now only accessible via FD.

One neat benefit of this (even when not using the gateway) is that now the sandcats / HTTPS code can be tested in dev mode.