Work around Npm.require() behavior change in Meteor 1.5.2.1.
Previously, Npm.require() had Node require()'s behavior where it searches up the directory tree for node_modules directories.

That was removed by: https://github.com/meteor/meteor/pull/9095

Sandstorm was relying on this, since the capnp module is built as part of the Sandstorm build and placed in an appropriate location where it can be found and used. We don't want to build capnp from npm since we want to use the version of libcapnp built as part of Sandstorm's own C++ build.

So, we now manually implement the behavior to find the capnp module and import it. This is encapsulated by a Meteor package which exports the global symbol `Capnp` for us.