Fix install script for less frequent releases
Reported via IRC. Sandstorm builds used to reliably occur weekly, so a month old installer was not surprising. As Sandstorm builds now happen infrequently, this causes a failure. As the tolerance here was four times the frequency of the average build, and we still regularly have new builds roughly monthly, I propose changing this to 120 days, or four times the current release cadence.