Bring Stripe-basaed payments and subscription code into Sandstorm.
This used to live in Blackrock only because we used to figure this code would be among the code we sell to hosting providers. But now Blackrock is open source, so this code might as well live in Sandstorm proper where anyone can use it. This also lets us remove a pile of hacks from the Blackrock codebase.