Clang 6 install instructions tweak.
I'm not sure what the appropriate amount of
detail for this should be.  We could walk the user through
adding the appropriate entries to their /etc/apt/sources.list.d/
directory, getting the repository signing key, and so forth.
How many users that reach this point in the instructions do not
know how to do that already?