Added a DNS note.
I set up Sandstorm on a VM for testing, and the local.sandstorm.io
address would not work for me.  It took me a few hours to track
down the problem, and it would have been longer if not for
TimMc on IRC.  My DD-WRT router was running dnsmasq and
blocking the resolution to 127.0.0.1.  So I've added a note
in the hope that nobody else stumbles on the same problem.