Fix large downloads / streams getting canceled after 2-3 minutes.
Along with the new Gateway code, grains now shut down if they've received no incoming calls in a particular 90-second period. This mostly works, but during a large download, a grain might do only outgoing calls, not incoming. So, the download could get cut off after 90-180 seconds. This also had the effect of breaking Groove Basin's audio streaming, which is based on a long download.

We don't want to trigger keep-alive on outgoing calls since it would be too easy for a grain to keep itself alive forever by accident. Grains should be required to register a wake lock if they want to stay running, so that the user is notified.

Instead, this change makes it so that WebSessionBridge periodically sends pings when performing a large download.

Fixes #3055