Modernize test suite.
The ancient version of Selenium we were using no longer works with the version of Firefox ESR on Debian Sid. Newer Seleniums no longer build-in Firefox support but rather require separately-installed drivers. Nightwatch also needed to be updated to work with newer Selenium. I switched to installing Selenium via npm as this seems more convenient.

For Firefox, the driver is `geckodriver`, but Nightwatch has a bug where its switchWindow() function doesn't work with `geckodriver`. The latest beta version of Nightwatch fixes this bug but seems to intorduce some other bugs. Instead, I switched to running the tests in Chrome.

Various tests needed updates to work correctly on the new version.