Remember identities through grain backup/restore.
Previously, when a grain was backed up and then restored, if the restored grain was shared with the same users as the original, the app would not know that they were the same users, as they would be assigned new, different identity IDs. After this change, identities will be remembered as long as the new user shares at least one of the original user's login credentials.